OT166-31: applied modifications suggested by the mentor
