+++ src/main/java/com/alkemy/ong/infrastructure/spring/config/security/JwtRequestFilter.java
@@ -1,23 +1,23 @@
 package com.alkemy.ong.infrastructure.spring.config.security;
 
+import com.alkemy.ong.infrastructure.util.JwtUtil;
 import java.io.IOException;
+import java.util.List;
+import java.util.stream.Collectors;
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 @Component
 public class JwtRequestFilter extends OncePerRequestFilter {
-
-  @Autowired
-  private UserDetailsCustomService userDetailsCustomService;
 
   @Autowired
   private JwtUtil jwtUtil;
@@ -30,18 +30,22 @@ public class JwtRequestFilter extends OncePerRequestFilter {
 
     String username = null;
     String jwt = null;
+    List<String> authorities = null;
 
-    if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+    if (jwtUtil.isTokenSet(authorizationHeader)) {
       jwt = authorizationHeader.substring(7);
       username = jwtUtil.extractUsername(jwt);
+      authorities = jwtUtil.getAuthorities(jwt);
     }
 
     if (username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
-      UserDetails userDetails = userDetailsCustomService.loadUserByUsername(username);
-      if (jwtUtil.validateToken(jwt, userDetails)) {
+
+      if (authorities != null) {
         UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken =
             new UsernamePasswordAuthenticationToken(
-            userDetails, null, userDetails.getAuthorities());
+                username,
+                "",
+                authorities.stream().map(SimpleGrantedAuthority::new).collect(Collectors.toList()));
 
         usernamePasswordAuthenticationToken.setDetails(
             new WebAuthenticationDetailsSource().buildDetails(request));
@@ -49,8 +53,6 @@ public class JwtRequestFilter extends OncePerRequestFilter {
         SecurityContextHolder.getContext().setAuthentication(usernamePasswordAuthenticationToken);
       }
     }
-
     filterChain.doFilter(request, response);
   }
-
 }
