--- src/main/java/com/alkemy/ong/infrastructure/spring/config/security/UserDetailsCustomService.java
@@ -1,17 +1,0 @@
-package com.alkemy.ong.infrastructure.spring.config.security;
-
-import java.util.ArrayList;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserDetailsCustomService implements UserDetailsService {
-
-  @Override
-  public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-    return new User("root", "pass123", new ArrayList<>());
-  }
-}
