+++ src/main/java/com/alkemy/ong/infrastructure/util/JwtUtil.java
@@ -1,0 +1,53 @@
+package com.alkemy.ong.infrastructure.util;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import java.nio.charset.StandardCharsets;
+import java.time.Instant;
+import java.util.Date;
+import java.util.List;
+import java.util.function.Function;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Service;
+
+@Service
+public class JwtUtil {
+
+  private static final String SECRET_KEY = "SECRET";
+
+  public String extractUsername(String token) {
+    return extractClaim(token, Claims::getSubject);
+  }
+
+  private <T> T extractClaim(String token, Function<Claims, T> claimsResolver) {
+    final Claims claims = extractAllClaims(token);
+    return claimsResolver.apply(claims);
+  }
+
+  private Claims extractAllClaims(String token) {
+    return Jwts.parser().setSigningKey(SECRET_KEY.getBytes(StandardCharsets.UTF_8))
+        .parseClaimsJws(token).getBody();
+  }
+
+  public String generateToken(UserDetails userDetails) {
+    return Jwts.builder()
+        .setSubject(userDetails.getUsername())
+        .claim("authorities",
+            userDetails.getAuthorities().stream().findFirst().get().getAuthority())
+        .setIssuedAt(Date.from(Instant.now()))
+        .signWith(SignatureAlgorithm.HS256, SECRET_KEY.getBytes(StandardCharsets.UTF_8)).compact();
+  }
+
+  public Boolean validateToken(String token, UserDetails userDetails) {
+    return extractUsername(token).equals(userDetails.getUsername());
+  }
+
+  public Boolean isTokenSet(String authorizationHeader) {
+    return authorizationHeader != null && authorizationHeader.startsWith("Bearer ");
+  }
+
+  public List<String> getAuthorities(String token) {
+    return (List<String>) extractAllClaims(token).get("authorities");
+  }
+}
